Rename, add docs
