+++ crates/zizmor/tests/integration/audit/insecure_url_scheme.rs
