+++ crates/zizmor/tests/integration/test-data/insecure-url-scheme/.pre-commit-config.yml
